chore: properly handle denied tool-invocation responses
