+++ platform/backend/src/routes/proxy/openai.ts
@@ -64,7 +64,7 @@ const openAiProxyRoutes: FastifyPluginAsyncZod = async (fastify) => {
 
       try {
         await utils.persistTools(tools, agentId);
-        await utils.evaluateTrustedDataPolicies(messages, chatId, agentId);
+        await utils.trustedData.evaluatePolicies(messages, chatId, agentId);
         await utils.persistUserMessage(messages, chatId);
 
         let assistantMessage: OpenAI.Chat.Completions.ChatCompletionMessage | null =
@@ -72,82 +72,18 @@ const openAiProxyRoutes: FastifyPluginAsyncZod = async (fastify) => {
 
         if (stream) {
           // Handle streaming response
-          reply.header("Content-Type", "text/event-stream");
-          reply.header("Cache-Control", "no-cache");
-          reply.header("Connection", "keep-alive");
-
           const stream = await openAiClient.chat.completions.create({
             ...body,
             stream: true,
           });
 
-          /**
-           * Accumulate the assistant message, and tool calls from chunks
-           *
-           * NOTE: for right now we ignore "custom" tool calls
-           */
-          let accumulatedContent = "";
-          const accumulatedToolCalls: OpenAI.Chat.Completions.ChatCompletionMessageFunctionToolCall[] =
-            [];
-
-          for await (const chunk of stream) {
-            const delta = chunk.choices[0]?.delta;
-
-            // Accumulate content
-            if (delta?.content) {
-              accumulatedContent += delta.content;
-            }
-
-            // Accumulate tool calls
-            if (delta?.tool_calls) {
-              for (const toolCallDelta of delta.tool_calls.filter(
-                (toolCall) => toolCall.type === "function",
-              )) {
-                const index = toolCallDelta.index;
-
-                // Initialize tool call if it doesn't exist
-                if (!accumulatedToolCalls[index]) {
-                  accumulatedToolCalls[index] = {
-                    id: toolCallDelta.id || "",
-                    type: "function",
-                    function: {
-                      name: "",
-                      arguments: "",
-                    },
-                  };
-                }
-
-                // Accumulate tool call fields
-                if (toolCallDelta.id) {
-                  accumulatedToolCalls[index].id = toolCallDelta.id;
-                }
-                if (toolCallDelta.function?.name) {
-                  accumulatedToolCalls[index].function.name =
-                    toolCallDelta.function.name;
-                }
-                if (toolCallDelta.function?.arguments) {
-                  accumulatedToolCalls[index].function.arguments +=
-                    toolCallDelta.function.arguments;
-                }
-              }
-            }
-
-            // Stream chunk to client
-            reply.raw.write(`data: ${JSON.stringify(chunk)}\n\n`);
-          }
-
-          // Construct the complete assistant message
-          assistantMessage = {
-            role: "assistant",
-            content: accumulatedContent || null,
-            tool_calls:
-              accumulatedToolCalls.length > 0
-                ? accumulatedToolCalls
-                : undefined,
-          } as OpenAI.Chat.Completions.ChatCompletionMessage;
+          const assistantMessage = await utils.streaming.handleChatCompletions(
+            reply,
+            stream,
+          );
 
           const toolInvocationPolicyError =
-            await utils.evaluateToolInvocationPolicies(
+            await utils.toolInvocation.evaluatePolicies(
               assistantMessage,
               agentId,
             );
@@ -177,7 +113,7 @@ const openAiProxyRoutes: FastifyPluginAsyncZod = async (fastify) => {
           assistantMessage = response.choices[0].message;
 
           const toolInvocationPolicyError =
-            await utils.evaluateToolInvocationPolicies(
+            await utils.toolInvocation.evaluatePolicies(
               assistantMessage,
               agentId,
             );
